Modificato metodo getTemperature per avere valori da 1 a 5

+++ garden-sensorboard/TemperatureSensor.h
@@ -5,7 +5,7 @@ class TemperatureSensor{
 
 public:
   TemperatureSensor(int pin);
-  float getTemperature();
+  int getTemperature();
 
 private:
   int pin;
